refactor: extract SyncStatusWidget and profile URL helpers from content.ts
Move SyncStatusWidget IIFE into lib/sync-status-widget.ts as a
DI factory (createSyncStatusWidget). Move buildSubmitMetadata,
getApiRowForIndex, isPlaceholderProfileUrl, extractProfileUrl
into the resume-extractor factory.

content.ts: 1630→1436 lines (-194)
