+++ apps/browser-extension/src/lib/sync-status-widget.ts
@@ -1,0 +1,162 @@
+// @ts-nocheck
+/**
+ * Sync status widget for browser extension content scripts.
+ * All dependencies injected from content.ts via DI factory.
+ */
+
+export function createSyncStatusWidget(deps) {
+  const {
+    win,
+    doc,
+    chrome,
+    onCancel,
+  } = deps;
+
+  const WIDGET_ID = "tr-sync-status-widget";
+  const DEFAULT_AUTO_DISMISS_MS = 5000;
+  const HIDE_DELAY_MS = 220;
+  let widgetEl = null;
+  let dismissTimer = null;
+  let hideTimer = null;
+
+  function escapeHtml(value) {
+    return String(value ?? "")
+      .replace(/&/g, "&amp;")
+      .replace(/</g, "&lt;")
+      .replace(/>/g, "&gt;")
+      .replace(/"/g, "&quot;")
+      .replace(/'/g, "&#39;");
+  }
+
+  function clearTimers() {
+    if (dismissTimer) {
+      clearTimeout(dismissTimer);
+      dismissTimer = null;
+    }
+    if (hideTimer) {
+      clearTimeout(hideTimer);
+      hideTimer = null;
+    }
+  }
+
+  function ensureWidget() {
+    if (widgetEl && widgetEl.isConnected) return widgetEl;
+
+    widgetEl = doc.createElement("div");
+    widgetEl.id = WIDGET_ID;
+    widgetEl.className = "tr-sync-widget";
+    widgetEl.setAttribute("role", "status");
+    widgetEl.setAttribute("aria-live", "polite");
+    widgetEl.setAttribute("aria-atomic", "true");
+    const mountTarget = doc.body || doc.documentElement;
+    mountTarget.appendChild(widgetEl);
+    return widgetEl;
+  }
+
+  function renderIcon(state) {
+    if (state === "progress") {
+      return '<span class="tr-sync-widget__spinner" aria-hidden="true"></span>';
+    }
+    if (state === "success") {
+      return '<span aria-hidden="true">✓</span>';
+    }
+    return '<span aria-hidden="true">!</span>';
+  }
+
+  function openOptionsPage() {
+    try {
+      void chrome.runtime
+        .sendMessage({ action: "openOptionsPage" })
+        .catch((error) => {
+          console.warn("🎯 [Auto Sync] Failed to open options page:", error);
+        });
+    } catch (error) {
+      console.warn("🎯 [Auto Sync] Failed to request options page:", error);
+    }
+  }
+
+  function show({
+    state = "progress",
+    message = "",
+    hint = "",
+    autoDismiss = false,
+  } = {}) {
+    const normalizedState =
+      state === "success" || state === "error" ? state : "progress";
+    const safeMessage = escapeHtml(message);
+    const safeHint = escapeHtml(hint);
+    const widget = ensureWidget();
+    clearTimers();
+
+    widget.className = `tr-sync-widget tr-sync-widget--${normalizedState}`;
+    widget.classList.remove("tr-sync-widget--hidden");
+    widget.innerHTML = `
+      <div class="tr-sync-widget__icon">${renderIcon(normalizedState)}</div>
+      <div class="tr-sync-widget__content">
+        <div class="tr-sync-widget__message">${safeMessage}</div>
+        ${safeHint ? `<div class="tr-sync-widget__hint">${safeHint}</div>` : ""}
+      </div>
+      ${
+        normalizedState === "progress"
+          ? '<button type="button" class="tr-sync-widget__cancel" aria-label="取消同步">取消</button>'
+          : normalizedState === "error"
+            ? '<button type="button" class="tr-sync-widget__close" aria-label="关闭提示">×</button>'
+            : ""
+      }
+    `;
+
+    widget.onclick = null;
+    if (normalizedState === "progress") {
+      const cancelBtn = widget.querySelector(".tr-sync-widget__cancel");
+      cancelBtn?.addEventListener("click", (event) => {
+        event.stopPropagation();
+        onCancel();
+        cancelBtn.setAttribute("disabled", "true");
+        cancelBtn.textContent = "取消中...";
+      });
+    }
+    if (normalizedState === "error") {
+      widget.onclick = (event) => {
+        const target = event.target instanceof Element ? event.target : null;
+        if (target?.closest(".tr-sync-widget__close")) return;
+        openOptionsPage();
+      };
+
+      const closeBtn = widget.querySelector(".tr-sync-widget__close");
+      closeBtn?.addEventListener("click", (event) => {
+        event.stopPropagation();
+        hide();
+      });
+    }
+
+    const dismissMs =
+      typeof autoDismiss === "number"
+        ? autoDismiss
+        : autoDismiss
+          ? DEFAULT_AUTO_DISMISS_MS
+          : 0;
+    if (dismissMs > 0) {
+      dismissTimer = setTimeout(() => {
+        hide();
+      }, dismissMs);
+    }
+  }
+
+  function hide() {
+    if (!widgetEl) return;
+    clearTimers();
+    widgetEl.classList.add("tr-sync-widget--hidden");
+    hideTimer = setTimeout(() => {
+      if (widgetEl) {
+        widgetEl.remove();
+        widgetEl = null;
+      }
+      hideTimer = null;
+    }, HIDE_DELAY_MS);
+  }
+
+  return {
+    show,
+    hide,
+  };
+}
